requirements.txt
psycopg2-binary
singlestoredb
snowflake-connector-python